requirements.txt
python-telegram-bot[webhooks]==20.7
gspread==5.8.0
google-auth==2.21.0